rename: 요청 객체와 응답 객체 패키지 분리
