+++ src/main/java/cart/presentation/CartItemApiController.java
@@ -1,9 +1,9 @@
 package cart.presentation;
 
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemAddRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.request.CartItemQuantityUpdateRequest;
+import cart.dto.request.CartItemAddRequest;
+import cart.dto.response.CartItemResponse;
 import cart.service.CartItemService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
