+++ src/main/java/cart/service/CartItemService.java
@@ -5,9 +5,9 @@ import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Product;
 import cart.domain.Quantity;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemAddRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.request.CartItemQuantityUpdateRequest;
+import cart.dto.request.CartItemAddRequest;
+import cart.dto.response.CartItemResponse;
 import cart.repository.CartItemRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
