+++ src/main/java/cart/service/OrderService.java
@@ -11,11 +11,11 @@ import cart.domain.Member;
 import cart.domain.OrderEntity;
 import cart.domain.OrderItemEntity;
 import cart.domain.Product;
-import cart.dto.CartItemInfoRequest;
-import cart.dto.OrderItemResponse;
-import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
-import cart.dto.ProductInfoRequest;
+import cart.dto.request.CartItemInfoRequest;
+import cart.dto.response.OrderItemResponse;
+import cart.dto.request.OrderRequest;
+import cart.dto.response.OrderResponse;
+import cart.dto.request.ProductInfoRequest;
 import cart.exception.MismatchedTotalPriceException;
 import cart.exception.MismatchedTotalProductPriceException;
 import cart.repository.CartItemRepository;
