+++ src/test/java/cart/integration/OrderIntegrationTest.java
@@ -8,9 +8,9 @@ import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Product;
 import cart.domain.Quantity;
-import cart.dto.CartItemInfoRequest;
-import cart.dto.OrderRequest;
-import cart.dto.ProductInfoRequest;
+import cart.dto.request.CartItemInfoRequest;
+import cart.dto.request.OrderRequest;
+import cart.dto.request.ProductInfoRequest;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
